requirements.txt
Flask>=2.2
requests==2.31.0
lxml
python-docx==1.1.0
python-dotenv==1.0.1
decipher
wheel
gunicorn

